--- Assets/Scripts/Player/Inventory/Consumable.cs
@@ -1,8 +1,0 @@
-﻿using UnityEngine;
-
-/// <summary> Расходник </summary>
-[CreateAssetMenu(fileName = "New consumable", menuName = "Items/Consumable")]
-public class Consumable : RenderableItem
-{
-
-}
