--- Assets/Scripts/Player/Inventory/Consumable.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e40339d6a652a3643a21b7aae3121f0d
