--- Assets/Scripts/Player/Inventory/IConsumable.cs
@@ -1,5 +1,0 @@
-﻿/// <summary> Расходник </summary>
-public interface IConsumable
-{
-    public void Use();
-}
