+++ Assets/Scripts/Player/Inventory/Inventory.cs
@@ -1,22 +1,27 @@
 ﻿using System.Collections.Generic;
+using Player.Items;
+using Player.Items.Implants;
 using UnityEngine;
 
-public class Inventory : MonoBehaviour
+namespace Player.Inventory
 {
-    [Header("Основные импланты")]
-    [SerializeField] public ArmsImplant armsImplant;
-    [SerializeField] public BodyImplant bodyImplant;
-    [SerializeField] public LegsImplant legsImplant;
+    public class Inventory : MonoBehaviour
+    {
+        [Header("Основные импланты")]
+        [SerializeField] private ArmsImplant armsImplant;
+        [SerializeField] private BodyImplant bodyImplant;
+        [SerializeField] private LegsImplant legsImplant;
 
-    [Header("Малые и гига-импланты")]
-    [SerializeField] public List<(SmallImplant, int)> smallImplants;
-    [SerializeField] public List<(GigaImplant, int)> gigaImplants;
+        [Header("Малые и гига-импланты")]
+        [SerializeField] private List<(SmallImplant, int)> smallImplants;
+        [SerializeField] private List<(GigaImplant, int)> gigaImplants;
 
-    [Header("Основное и второстепенное оружие")]
-    [SerializeField] public GunInfo primaryWeapon;
-    [SerializeField] public GunInfo secondaryWeapon;
+        [Header("Основное и второстепенное оружие")]
+        [SerializeField] private GunInfo primaryWeapon;
+        [SerializeField] private GunInfo secondaryWeapon;
 
-    [Header("Расходники")]
-    [SerializeField] public IConsumable[] craftedConsumables = new IConsumable[3];
-    [SerializeField] public List<(IConsumable, int)> consumables;
+        [Header("Расходники")]
+        [SerializeField] private IConsumable[] craftedConsumables = new IConsumable[3];
+        [SerializeField] private List<(IConsumable, int)> consumables;
+    }
 }
