--- Assets/Scripts/Player/Inventory/LegsImplant.cs
@@ -1,8 +1,0 @@
-﻿using UnityEngine;
-
-/// <summary> Имплант первого слоя под ноги </summary>
-//[CreateAssetMenu(fileName = "New legs implant", menuName = "Items/Implants/Legs")]
-public abstract class LegsImplant : Implant
-{
-    public abstract void ImplantAction();
-}
