--- Assets/Scripts/Player/Inventory/RenderableItem.cs
@@ -1,9 +1,0 @@
-﻿using UnityEngine;
-
-/// <summary> Предмет с 3D моделью </summary>
-public abstract class RenderableItem : Item
-{
-    [Header("Визуализация")]
-    /// <summary> 3D модель предмета </summary>
-    public GameObject prefab;
-}
