--- Assets/Scripts/Player/Inventory/SmallImplant.cs
@@ -1,9 +1,0 @@
-﻿using UnityEngine;
-
-/// <summary> Малый мплант второго слоя </summary>
-[CreateAssetMenu(fileName = "New small implant", menuName = "Items/Implants/Small")]
-public abstract class SmallImplant : Implant
-{
-    /// <summary> Редкость импланта </summary>
-    public Rareness type = Rareness.Common;
-}
