+++ Assets/Scripts/Player/Items.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3111913020be7904396e0ed912cb57d3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
