+++ Assets/Scripts/Player/Items/IConsumable.cs
@@ -1,0 +1,7 @@
+﻿namespace Player.Items
+{
+    public interface IConsumable
+    {
+        public void Use();
+    }
+}
