+++ Assets/Scripts/Player/Items/IConsumable.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: a7e58c8ec7822b6429f195619c283ff3
