+++ Assets/Scripts/Player/Items/Implants.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7f9e9fa5b0cad1c499ccffcff99f03b4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
