+++ Assets/Scripts/Player/Items/Implants/ArmsImplant.cs
@@ -1,0 +1,7 @@
+﻿namespace Player.Items.Implants
+{
+    /// <summary> Имплант первого слоя под руки </summary>
+    public abstract class ArmsImplant : Implant
+    {
+    }
+}
