+++ Assets/Scripts/Player/Items/Implants/BodyImplant.cs
@@ -1,0 +1,7 @@
+﻿namespace Player.Items.Implants
+{
+    /// <summary> Имплант первого слоя под тело </summary>
+    public abstract class BodyImplant : Implant
+    {
+    }
+}
