+++ Assets/Scripts/Player/Items/Implants/GigaImplant.cs
@@ -1,0 +1,8 @@
+﻿namespace Player.Items.Implants
+{
+    /// <summary> Гига-имплант третьего слоя </summary>
+    public abstract class GigaImplant : Implant
+    {
+        public Rareness type = Rareness.Common;
+    }
+}
