+++ Assets/Scripts/Player/Items/Implants/IImplantable.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace Player.Items.Implants
+{
+    public interface IImplantable
+    {
+        public void Action();
+    }
+}
