+++ Assets/Scripts/Player/Items/Implants/IImplantable.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 9e06ae4f2f224704c8f21b082723a2ac
