+++ Assets/Scripts/Player/Items/Implants/Implant.cs
@@ -1,0 +1,9 @@
+﻿namespace Player.Items.Implants
+{
+    public abstract class Implant : Item, IImplantable
+    {
+        public enum Rareness { Common, Rare }
+
+        public abstract void Action();
+    }
+}
