+++ Assets/Scripts/Player/Items/Implants/LegsImplant.cs
@@ -1,0 +1,7 @@
+﻿namespace Player.Items.Implants
+{
+    /// <summary> Имплант первого слоя под ноги </summary>
+    public abstract class LegsImplant : Implant
+    {
+    }
+}
