+++ Assets/Scripts/Player/Items/Implants/LegsImplant.cs.meta
