+++ Assets/Scripts/Player/Items/Implants/SmallImplant.cs
@@ -1,0 +1,8 @@
+﻿namespace Player.Items.Implants
+{
+    /// <summary> Малый мплант второго слоя </summary>
+    public abstract class SmallImplant : Implant
+    {
+        public Rareness type = Rareness.Common;
+    }
+}
