+++ Assets/Scripts/Player/Items/RenderableItem.cs
@@ -1,0 +1,11 @@
+﻿using UnityEngine;
+
+namespace Player.Items
+{
+    /// <summary> Предмет с 3D моделью </summary>
+    public abstract class RenderableItem : Item
+    {
+        [Header("Визуализация")]
+        public GameObject prefab;
+    }
+}
