+++ Assets/Scripts/Player/Items/RenderableItem.cs.meta
