--- Assets/Scripts/Player/PlayerControllerRebound42.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: b0db8c9590169c84f96ec60f96ed3746
