+++ Assets/Scripts/Weapon/WeaponInputHandler.cs
@@ -33,7 +33,7 @@ public class WeaponInputHandler : MonoBehaviour
         }
         else
         {
-            Debug.Log($"[WeaponInputHandler] Найдено оружие: {weapon.gunInfo.gunName}");
+            Debug.Log($"[WeaponInputHandler] Найдено оружие: {weapon.gunInfo.name}");
             Debug.Log("ЛКМ - стрелять/атаковать, R - перезарядка");
         }
     }
@@ -93,7 +93,7 @@ public class WeaponInputHandler : MonoBehaviour
         style.normal.textColor = Color.white;
 
         GUI.Box(new Rect(10, 10, 250, 120), "");
-        GUI.Label(new Rect(20, 15, 300, 30), $"Оружие: {weapon.gunInfo.gunName}", style);
+        GUI.Label(new Rect(20, 15, 300, 30), $"Оружие: {weapon.gunInfo.name}", style);
 
         if (weapon.WeaponType == GunInfo.WeaponType.Firearm)
         {
